+++ molo/surveys/tests/test_views.py
@@ -69,7 +69,8 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
     def create_molo_survey_page(self, parent, **kwargs):
         molo_survey_page = MoloSurveyPage(
             title='Test Survey', slug='test-survey',
-            intro='Introduction to Test Survey ...',
+            introduction='Introduction to Test Survey ...',
+            homepage_introduction='Shorter homepage introduction',
             thank_you_text='Thank you for taking the Test Survey',
             submit_text='survey submission text',
             **kwargs
@@ -96,6 +97,16 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         self.assertContains(response, 'share your story yo')
         self.assertNotContains(response, 'Take the Survey')
 
+    def test_correct_intro_shows_on_homepage(self):
+        molo_survey_page, molo_survey_form_field = \
+            self.create_molo_survey_page(
+                parent=self.surveys_index,
+                homepage_button_text='share your story yo')
+        self.client.login(username='tester', password='tester')
+        response = self.client.get('/')
+        self.assertContains(response, 'Shorter homepage introduction')
+        self.assertNotContains(response, 'Take the Survey')
+
     def test_anonymous_submissions_not_allowed_by_default(self):
         molo_survey_page, molo_survey_form_field = \
             self.create_molo_survey_page(parent=self.section_index)
@@ -113,7 +124,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.get(molo_survey_page.url)
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
         self.assertContains(response, molo_survey_page.submit_text)
 
@@ -136,7 +147,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client.get(molo_survey_page.url)
 
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
         response = self.client.post(molo_survey_page.url, {
             molo_survey_form_field.label.lower().replace(' ', '-'): 'python'
@@ -177,7 +188,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
             response = self.client.get(molo_survey_page.url)
 
             self.assertContains(response, molo_survey_page.title)
-            self.assertContains(response, molo_survey_page.intro)
+            self.assertContains(response, molo_survey_page.introduction)
             self.assertContains(response, molo_survey_form_field.label)
 
             response = self.client.post(molo_survey_page.url, {
@@ -200,7 +211,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.get(molo_survey_page.url)
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
 
         response = self.client.post(molo_survey_page.url, {
@@ -223,7 +234,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.get(molo_survey_page.url)
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
 
         response = self.client.post(molo_survey_page.url, {
@@ -261,7 +272,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client.get(molo_survey_page.url)
 
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
         self.assertNotContains(response, extra_molo_survey_form_field.label)
         self.assertContains(response, 'Next Question')
@@ -271,7 +282,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         })
 
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertNotContains(response, molo_survey_form_field.label)
         self.assertContains(response, extra_molo_survey_form_field.label)
         self.assertContains(response, molo_survey_page.submit_text)
@@ -296,7 +307,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         response = self.client.get(molo_survey_page.url)
 
         self.assertContains(response, molo_survey_page.title)
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.introduction)
         self.assertContains(response, molo_survey_form_field.label)
 
         response = self.client.post(molo_survey_page.url, {})
@@ -326,7 +337,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
             self.create_molo_survey_page(parent=self.surveys_index)
         response = self.client.get("/")
         self.assertContains(response, 'Take The Survey</a>')
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.homepage_introduction)
         user = User.objects.create_superuser(
             username='testuser', password='password', email='test@email.com')
         self.client2.login(user=user)
@@ -338,7 +349,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
             self.create_molo_survey_page(parent=self.surveys_index)
         response = self.client.get("/")
         self.assertContains(response, 'Take The Survey</a>')
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.homepage_introduction)
         user = User.objects.create_superuser(
             username='testuser', password='password', email='test@email.com')
         self.client2.login(user=user)
@@ -443,7 +454,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
 
         response = self.client.get(self.section.url)
         self.assertContains(response, 'Take The Survey</a>')
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.homepage_introduction)
 
     def test_translated_survey_on_section_page(self):
         self.user = self.login()
@@ -481,7 +492,7 @@ class TestSurveyViews(TestCase, MoloTestCaseMixin):
         self.assertContains(response,
                             'Take The Survey</a>'.format(
                                 molo_survey_page.url))
-        self.assertContains(response, molo_survey_page.intro)
+        self.assertContains(response, molo_survey_page.homepage_introduction)
 
     def test_survey_list_display_direct_logged_out(self):
         molo_survey_page, molo_survey_form_field = \
@@ -679,7 +690,7 @@ class TestSkipLogicSurveyView(TestCase, MoloTestCaseMixin):
     def new_survey(self, name):
         survey = MoloSurveyPage(
             title=name, slug=slugify(name),
-            intro='Introduction to {}...'.format(name),
+            introduction='Introduction to {}...'.format(name),
             thank_you_text='Thank you for taking the {}'.format(name),
             submit_text='survey submission text for {}'.format(name),
             allow_anonymous_submissions=True,
@@ -690,7 +701,7 @@ class TestSkipLogicSurveyView(TestCase, MoloTestCaseMixin):
 
     def assertSurveyAndQuestions(self, response, survey, questions):
         self.assertContains(response, survey.title)
-        self.assertContains(response, survey.intro)
+        self.assertContains(response, survey.introduction)
         for question in questions:
             self.assertContains(response, question.label)
             self.assertContains(response, question.label)